removes all (almost) all dirty frontend magic
